fixes Respawn transition ,Respawn Trigger
