+++ Where_Is_My_Wife/Assets/_Main/Assets/Stage/Animations/RespawnCurtainsAnimation.cs
@@ -7,51 +7,60 @@ using WhereIsMyWife.Managers;
 
 namespace WhereIsMyWife.Controllers
 {
-        public class RespawnCurtainsAnimation : MonoBehaviour
+    public class RespawnCurtainsAnimation : MonoBehaviour
+    {
+        [SerializeField] private Image _leftPanel;
+        [SerializeField] private Image _rightPanel;
+        [SerializeField] private float _animationDuration;
+        [SerializeField] private float _closedTime;
+        [SerializeField] private Ease _ease = Ease.InSine;
+
+        private float _panelWidth;
+
+        private void Awake()
         {
-            [SerializeField] private Image _leftPanel;
-            [SerializeField] private Image _rightPanel;
-            [SerializeField] private float _animationDuration;
-            [SerializeField] private float _closedTime;
-            [SerializeField] private Ease _ease = Ease.InSine;
+            _panelWidth = _leftPanel.rectTransform.rect.width;
+        }
 
-            private float _panelWidth;
+        private void Start()
+        {
+            PlayerManager.Instance.DeathAction += StartAnimation;
+        }
 
-            private void Awake()
-            {
-                _panelWidth = _leftPanel.rectTransform.rect.width;
-            }
+        private void OnDestroy()
+        {
+            PlayerManager.Instance.DeathAction -= StartAnimation;
+        }
 
-            private void Start()
-            {
-                PlayerManager.Instance.DeathAction += StartAnimation;
-            }
+        private void StartAnimation()
+        {
+            AnimationAsync().Forget();
+        }
 
-            private void OnDestroy()
-            {
-                PlayerManager.Instance.DeathAction -= StartAnimation;
-            }
+        private async UniTaskVoid AnimationAsync()
+        {
+            _leftPanel.rectTransform.DOAnchorPosX(_panelWidth, _animationDuration).SetEase(_ease);
+            await _rightPanel.rectTransform.DOAnchorPosX(-_panelWidth, _animationDuration)
+                .SetEase(_ease).AsyncWaitForCompletion();
 
-            private void StartAnimation()
-            {
-                AnimationAsync().Forget();
-            }
+            await UniTask.Delay(TimeSpan.FromSeconds(_closedTime));
 
-            private async UniTaskVoid AnimationAsync()
-            {
-                _leftPanel.rectTransform.DOAnchorPosX(_panelWidth, _animationDuration).SetEase(_ease);
-                await _rightPanel.rectTransform.DOAnchorPosX(-_panelWidth, _animationDuration)
-                    .SetEase(_ease).AsyncWaitForCompletion();
+            PlayerManager.Instance.TriggerRespawnStart();
 
-                await UniTask.Delay(TimeSpan.FromSeconds(_closedTime));
+            _leftPanel.rectTransform.DOAnchorPosX(0, _animationDuration).SetEase(_ease);
+            await _rightPanel.rectTransform.DOAnchorPosX(0, _animationDuration)
+                .SetEase(_ease).AsyncWaitForCompletion();
 
-                PlayerManager.Instance.TriggerRespawnStart();
-
-                _leftPanel.rectTransform.DOAnchorPosX(0, _animationDuration).SetEase(_ease);
-                await _rightPanel.rectTransform.DOAnchorPosX(0, _animationDuration)
-                    .SetEase(_ease).AsyncWaitForCompletion();
-
-                PlayerManager.Instance.TriggerRespawnComplete();
-            }
+            await RespawnPointTransitionAnimation.RespawnComplete();
         }
     }
+
+    public static class RespawnPointTransitionAnimation
+    {
+        public static async UniTask RespawnComplete()
+        {
+            PlayerManager.Instance.TriggerRespawnComplete();
+            await UniTask.Yield();  
+        }
+    }
+}
