+++ Where_Is_My_Wife/Assets/_Main/Assets/Stage/Animations/RespawnTrigger.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 26e65009071c9844b94fc424bb8daa79
+guid: e7a0459c936b60640843cdd5153e4b03
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
