+++ Where_Is_My_Wife/Assets/_Main/Scenes/Dash_Test.unity
@@ -5578,7 +5578,7 @@ Transform:
   m_GameObject: {fileID: 1759852571}
   serializedVersion: 2
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 113.44, y: 2.98, z: 0}
+  m_LocalPosition: {x: 212.9, y: 1, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
